requirements.txt
fastapi
uvicorn[standard]
websockets
langgraph
langchain-ollama
langchain-core
httpx
pydantic
twilio
python-dotenv